requirements.txt
psycopg2
djangorestframework
django-cors-headers
dotenv
djangorestframework-simplejwt
django-filter
# pip install -r requirements.txt